requirements.txt
requests
pandas
scikit-learn
apache-airflow
boto3
fastavro
PyYAML